Move Apply Changes button to ShellPage to deduplicate across settings pages
