+++ src/windows/wslsettings/Views/Settings/MemAndProcPage.xaml.cs
@@ -43,7 +43,6 @@ public sealed partial class MemAndProcPage : Page
     override protected void OnNavigatedFrom(NavigationEventArgs e)
     {
         App.GetService<IWslConfigService>().WslConfigChanged -= ViewModel.OnConfigChanged;
-        App.GetService<IWslConfigService>().PendingChangesChanged -= ViewModel.OnPendingChangesChanged;
     }
 
     private void Settings_ResetButton_Click(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
@@ -97,10 +96,5 @@ public sealed partial class MemAndProcPage : Page
 
         TextBox? textBox = sender as TextBox;
         ViewModel.SetSwapSize_ResetEnabled(textBox!.Text);
-    }
-
-    private async void ApplyChanges_Click(object sender, RoutedEventArgs e)
-    {
-        await SettingsApplyHelper.ShowApplyChangesDialogAsync(XamlRoot);
     }
 }
