+++ src/windows/wslsettings/Views/Settings/NetworkingPage.xaml.cs
@@ -41,7 +41,6 @@ public sealed partial class NetworkingPage : Page
     override protected void OnNavigatedFrom(NavigationEventArgs e)
     {
         App.GetService<IWslConfigService>().WslConfigChanged -= ViewModel.OnConfigChanged;
-        App.GetService<IWslConfigService>().PendingChangesChanged -= ViewModel.OnPendingChangesChanged;
     }
 
     private void Settings_ResetButton_Click(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
@@ -63,10 +62,5 @@ public sealed partial class NetworkingPage : Page
 
         TextBox? textBox = sender as TextBox;
         ViewModel.SetInitialAutoProxyTimeout_ResetEnabled(textBox!.Text);
-    }
-
-    private async void ApplyChanges_Click(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
-    {
-        await SettingsApplyHelper.ShowApplyChangesDialogAsync(XamlRoot);
     }
 }
