+++ src/windows/wslsettings/Views/Settings/OptionalFeaturesPage.xaml.cs
@@ -40,7 +40,6 @@ public sealed partial class OptionalFeaturesPage : Page
     override protected void OnNavigatedFrom(NavigationEventArgs e)
     {
         App.GetService<IWslConfigService>().WslConfigChanged -= ViewModel.OnConfigChanged;
-        App.GetService<IWslConfigService>().PendingChangesChanged -= ViewModel.OnPendingChangesChanged;
     }
 
     private void Settings_ResetButton_Click(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
@@ -62,10 +61,5 @@ public sealed partial class OptionalFeaturesPage : Page
 
         TextBox? textBox = sender as TextBox;
         ViewModel.SetVMIdleTimeout_ResetEnabled(textBox!.Text);
-    }
-
-    private async void ApplyChanges_Click(object sender, Microsoft.UI.Xaml.RoutedEventArgs e)
-    {
-        await SettingsApplyHelper.ShowApplyChangesDialogAsync(XamlRoot);
     }
 }
